+++ lib/screens/home/library_tab/routine/routine_detail_screen.dart
@@ -1,8 +1,9 @@
 import 'package:cached_network_image/cached_network_image.dart';
+import 'package:feature_discovery/feature_discovery.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
+import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:provider/provider.dart' as provider;
@@ -86,6 +87,14 @@ class _RoutineDetailScreenState extends State<RoutineDetailScreen>
   void initState() {
     super.initState();
     widget.model.init(this);
+    SchedulerBinding.instance!.addPostFrameCallback((Duration duration) {
+      FeatureDiscovery.discoverFeatures(
+        context,
+        const <String>{
+          'reorder_routine_workouts',
+        },
+      );
+    });
   }
 
   @override